Display polyline reflecting real route
Accessed the Google Maps Directions API, which sends back an array of coordinates based on a starting and ending city.  Get request returns an array of coordinates which can then be used to display an accurate polyline - previously, the polyline was simply a straight line from starting to ending city.