Move route render functionality into MappedRoute
Moved functionality that renders the route from Results to MappedRoute component.  Created function renderRoute to hold functionality, and invoke function once user enters a commute search and startCoords property is passed down from Results.  This ensures the route will be rendered once a search is entered and the bounds of the map will center and zoom accordingly.